# fix: use gather for get_proj_info