This seems to speed things up for me pretty dramatically. Some very naive timing just with my stopwatch on my phone it goes like this:

main:
- Test output starting to show: 20 seconds
- Finished: 34 seconds

This branch:
- Test output starting to show: 6 seconds
- Finished: 20 seconds

So, in my project it reduces the time down from 20 to 6 seconds.

The last 14 seconds is building the projects, and finding the tests.

From my initial testing which I also mentioned in the issue, it seems like building is ~12 of these 14 seconds