@Nsidorenco 2025-06-24
Looks good now. Let's get this blazingly fast optimization into main 🚀